fix: daily-paymentの反映
