--- functions/src/daily-usages/withdraw.ts
@@ -1,235 +1,0 @@
-/* eslint-disable @typescript-eslint/no-var-requires */
-
-/* eslint-disable camelcase */
-import { daily_usage } from '.';
-import { account_private } from '../account-privates';
-import { admin_account } from '../admin-accounts';
-import { balance } from '../balances';
-import { daily_payment } from '../daily-payments';
-import { insufficient_balance } from '../insufficient-balances';
-import { student_account } from '../student-accounts';
-import { InsufficientBalance, DailyPayment, DailyUsage, Balance } from '@local/common';
-import * as crypto from 'crypto-js';
-import * as functions from 'firebase-functions';
-
-daily_usage.onCreateHandler.push(async (snapshot, context) => {
-  const data = snapshot.data()! as DailyUsage;
-  const usage = parseInt(data.amount_kwh_str) * 1000000;
-  const studentAccounts = await student_account.getByRoomID(data.room_id);
-  if (!studentAccounts.length) {
-    console.log(data.room_id, 'no student');
-    return;
-  }
-  if (usage <= 0) {
-    console.log('0 or minus usage detected');
-    return;
-  }
-  for (const student of studentAccounts) {
-    const accountBalance = await balance.listLatest(student.id);
-    const uupxAmount = parseInt(accountBalance[0].amount_uupx);
-    const uspxAmount = parseInt(accountBalance[0].amount_uspx);
-    const totalBalance = uupxAmount + uspxAmount;
-    const accountPrivate = await account_private.list(student.id);
-    const xrpl = require('xrpl');
-    const TEST_NET = 'wss://s.altnet.rippletest.net:51233';
-    const client = new xrpl.Client(TEST_NET);
-    const adminAccount = await admin_account.getByName('admin');
-    const date = new Date();
-
-    await daily_payment.create(
-      new DailyPayment({
-        student_account_id: student.id,
-        year: date.getFullYear().toString(),
-        month: (date.getMonth() + 1).toString(),
-        date: date.getDate().toString(),
-        amount_mwh: usage.toString(),
-      }),
-    );
-
-    if (usage <= uspxAmount) {
-      await balance.create(
-        new Balance({
-          id: accountBalance[0].id,
-          student_account_id: accountBalance[0].student_account_id,
-          amount_uupx: accountBalance[0].amount_uupx,
-          amount_uspx: (uspxAmount - usage).toString(),
-        }),
-      );
-      if (!accountPrivate.length) {
-        console.log(student.id, 'no XRP address');
-        return;
-      }
-      await client.connect();
-      const config = functions.config();
-      const confXrpl = config['xrpl'];
-      const privKey = confXrpl.private_key;
-      const decrypted = crypto.AES.decrypt(accountPrivate[0].xrp_seed, privKey).toString(crypto.enc.Utf8);
-      const sender = xrpl.Wallet.fromSeed(decrypted);
-      const vli = await client.getLedgerIndex();
-
-      const sendTokenTx = {
-        TransactionType: 'Payment',
-        Account: sender.address,
-        Amount: {
-          currency: 'SPX',
-          value: (usage / 1000000).toString(),
-          issuer: adminAccount[0].xrp_address_cold,
-        },
-        Destination: adminAccount[0].xrp_address_hot,
-        LastLedgerSequence: vli + 150,
-      };
-      const payPrepared = await client.autofill(sendTokenTx);
-      const paySigned = sender.sign(payPrepared);
-      const payResult = await client.submitAndWait(paySigned.tx_blob);
-      if (payResult.result.meta.TransactionResult == 'tesSUCCESS') {
-        console.log(`Transaction succeeded: https://testnet.xrpl.org/transactions/${paySigned.hash}`);
-      } else {
-        // eslint-disable-next-line no-throw-literal
-        throw `Error sending transaction: ${payResult.result.meta.TransactionResult}`;
-      }
-      client.disconnect();
-    } else if (usage < totalBalance) {
-      const uspxShortage = usage - uspxAmount;
-      await balance.create(
-        new Balance({
-          id: accountBalance[0].id,
-          student_account_id: accountBalance[0].student_account_id,
-          amount_uupx: (uupxAmount - uspxShortage).toString(),
-          amount_uspx: '0',
-        }),
-      );
-      if (!accountPrivate.length) {
-        console.log(student.id, 'no XRP address');
-        return;
-      }
-      await client.connect();
-
-      const config = functions.config();
-      const confXrpl = config['xrpl'];
-      const privKey = confXrpl.private_key;
-      const decrypted = crypto.AES.decrypt(accountPrivate[0].xrp_seed, privKey).toString(crypto.enc.Utf8);
-      const sender = xrpl.Wallet.fromSeed(decrypted);
-      let vli = await client.getLedgerIndex();
-
-      if (uspxAmount > 0) {
-        const sendSPXTx = {
-          TransactionType: 'Payment',
-          Account: sender.address,
-          Amount: {
-            currency: 'SPX',
-            value: (uspxAmount / 1000000).toString(),
-            issuer: adminAccount[0].xrp_address_cold,
-          },
-          Destination: adminAccount[0].xrp_address_hot,
-          LastLedgerSequence: vli + 150,
-        };
-        const payPreparedSPX = await client.autofill(sendSPXTx);
-        const paySignedSPX = sender.sign(payPreparedSPX);
-        const payResultSPX = await client.submitAndWait(paySignedSPX.tx_blob);
-        if (payResultSPX.result.meta.TransactionResult == 'tesSUCCESS') {
-          console.log(`Transaction succeeded: https://testnet.xrpl.org/transactions/${paySignedSPX.hash}`);
-        } else {
-          // eslint-disable-next-line no-throw-literal
-          throw `Error sending transaction: ${payResultSPX.result.meta.TransactionResult}`;
-        }
-      }
-      vli = await client.getLedgerIndex();
-
-      const sendUPXTx = {
-        TransactionType: 'Payment',
-        Account: sender.address,
-        Amount: {
-          currency: 'UPX',
-          value: (uspxShortage / 1000000).toString(),
-          issuer: adminAccount[0].xrp_address_cold,
-        },
-        Destination: adminAccount[0].xrp_address_hot,
-        LastLedgerSequence: vli + 150,
-      };
-      const payPreparedUPX = await client.autofill(sendUPXTx);
-      const paySignedUPX = sender.sign(payPreparedUPX);
-      const payResultUPX = await client.submitAndWait(paySignedUPX.tx_blob);
-      if (payResultUPX.result.meta.TransactionResult == 'tesSUCCESS') {
-        console.log(`Transaction succeeded: https://testnet.xrpl.org/transactions/${paySignedUPX.hash}`);
-      } else {
-        // eslint-disable-next-line no-throw-literal
-        throw `Error sending transaction: ${payResultUPX.result.meta.TransactionResult}`;
-      }
-      client.disconnect();
-    } else {
-      const insufficiency = usage - totalBalance;
-      console.log('不足分', insufficiency);
-      await balance.create(
-        new Balance({
-          id: accountBalance[0].id,
-          student_account_id: accountBalance[0].student_account_id,
-          amount_uupx: '0',
-          amount_uspx: '0',
-        }),
-      );
-      await insufficient_balance.create(
-        new InsufficientBalance({ student_account_id: accountBalance[0].student_account_id, amount_utoken: insufficiency.toString() }),
-      );
-
-      if (!accountPrivate.length) {
-        console.log(student.id, 'no XRP address');
-        return;
-      }
-      await client.connect();
-      const config = functions.config();
-      const confXrpl = config['xrpl'];
-      const privKey = confXrpl.private_key;
-      const decrypted = crypto.AES.decrypt(accountPrivate[0].xrp_seed, privKey).toString(crypto.enc.Utf8);
-      const sender = xrpl.Wallet.fromSeed(decrypted);
-      let vli = await client.getLedgerIndex();
-
-      if (uspxAmount > 0) {
-        const sendSPXTx = {
-          TransactionType: 'Payment',
-          Account: sender.address,
-          Amount: {
-            currency: 'SPX',
-            value: (uspxAmount / 1000000).toString(),
-            issuer: adminAccount[0].xrp_address_cold,
-          },
-          Destination: adminAccount[0].xrp_address_hot,
-          LastLedgerSequence: vli + 150,
-        };
-        const payPreparedSPX = await client.autofill(sendSPXTx);
-        const paySignedSPX = sender.sign(payPreparedSPX);
-        const payResultSPX = await client.submitAndWait(paySignedSPX.tx_blob);
-        if (payResultSPX.result.meta.TransactionResult == 'tesSUCCESS') {
-          console.log(`Transaction succeeded: https://testnet.xrpl.org/transactions/${paySignedSPX.hash}`);
-        } else {
-          // eslint-disable-next-line no-throw-literal
-          throw `Error sending transaction: ${payResultSPX.result.meta.TransactionResult}`;
-        }
-      }
-      vli = await client.getLedgerIndex();
-
-      if (uupxAmount > 0) {
-        const sendUPXTx = {
-          TransactionType: 'Payment',
-          Account: sender.address,
-          Amount: {
-            currency: 'UPX',
-            value: (uupxAmount / 1000000).toString(),
-            issuer: adminAccount[0].xrp_address_cold,
-          },
-          Destination: adminAccount[0].xrp_address_hot,
-          LastLedgerSequence: vli + 150,
-        };
-        const payPreparedUPX = await client.autofill(sendUPXTx);
-        const paySignedUPX = sender.sign(payPreparedUPX);
-        const payResultUPX = await client.submitAndWait(paySignedUPX.tx_blob);
-        if (payResultUPX.result.meta.TransactionResult == 'tesSUCCESS') {
-          console.log(`Transaction succeeded: https://testnet.xrpl.org/transactions/${paySignedUPX.hash}`);
-        } else {
-          // eslint-disable-next-line no-throw-literal
-          throw `Error sending transaction: ${payResultUPX.result.meta.TransactionResult}`;
-        }
-      }
-      client.disconnect();
-    }
-  }
-});
